WIP: run all tests against helm CO
